Extract MailSender interface
